Added support for sample info file in same directory
Sample info files retrieved by the .bat file can now be used in the program. It will first look for this file. If it does not exist it will look for the sample info file in the .D folder. If both do not exist, the sample info will be set as empty strings.