Fixed bug where non-cocaine spectra could be seen as cocaine-containing
The algorithm sometimes misinterpreted a peak as a cocaine peak and calculated a wrong RT modifier. This resulted in the wrong determination of the internal standard and could result in a incorrect integration and cocaine-IS ratio.
Fixed this by now also calculation the RT modifier based on the IS. The RT modifier of the IS and cocaine are then compared and cannot differ more than 0.015